Move logger to src directory
